Moved VRInterface outside GazeRetargeting
