+++ modules/SRanipal_module/CMakeLists.txt
@@ -12,13 +12,15 @@ set (${EXE_TARGET_NAME}_SRC src/main.cpp
                             src/SRanipalInterface.cpp
                             src/EyelidsRetargeting.cpp
                             src/FaceExpressionsRetargeting.cpp
-                            src/GazeRetargeting.cpp)
+                            src/GazeRetargeting.cpp
+                            src/VRInterface.cpp)
 
 set (${EXE_TARGET_NAME}_HDR include/SRanipalModule.hpp
                             include/SRanipalInterface.hpp
                             include/EyelidsRetargeting.hpp
                             include/FaceExpressionsRetargeting.hpp
-                            include/GazeRetargeting.hpp)
+                            include/GazeRetargeting.hpp
+                            include/VRInterface.hpp)
 
 add_executable(${EXE_TARGET_NAME} ${${EXE_TARGET_NAME}_SRC} ${${EXE_TARGET_NAME}_HDR})
 target_link_libraries(${EXE_TARGET_NAME} PRIVATE ${YARP_LIBRARIES} Eigen3::Eigen iDynTree::idyntree-core SRanipalSDK::SRanipalSDK)
